curriculum-tracker/migrations/versions/a9570b3d9429_add_users_table.py
"""add_users_table

Revision ID: a9570b3d9429
Revises: 844f95627dee
Create Date: 2025-12-01 19:56:01.409753

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = 'a9570b3d9429'
down_revision: Union[str, Sequence[str], None] = '844f95627dee'
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    """Upgrade schema."""
    op.execute("""
        CREATE TABLE users (
            id SERIAL PRIMARY KEY,
            username TEXT UNIQUE NOT NULL,
            password_hash TEXT NOT NULL,
            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
        )
    """)


def downgrade() -> None:
    """Downgrade schema."""
    op.execute("DROP TABLE IF EXISTS users CASCADE")
