curriculum-tracker/requirements.txt
click>=8.0.0
pyyaml>=6.0
rich>=13.0.0
flask>=3.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
alembic>=1.13.0
Flask-Login>=0.6.3
werkzeug>=3.0.0
pytest>=7.4.0
pytest-flask>=1.3.0
